refactor (transfer creating result responsibility to winningLotto)
